requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
peewee==3.14.4
python-dotenv==0.17.1
Werkzeug==1.0.1
